data is persistent now

+++ BusShuttleProject/BusShuttle/FileSaver.cs
@@ -5,10 +5,13 @@ using System.IO;
 public class FileSaver
 {
     string fileName;
+
     public FileSaver(string fileName)
     {
         this.fileName = fileName;
-        File.Create(this.fileName).Close();
+        if(!File.Exists(this.fileName)) {
+            File.Create(this.fileName).Close();
+        }
     }
     public void AppendLine(string line)
     {
